Move MultiRowProcessorType to new file
