+++ src/main/java/org/nkjmlab/sorm4j/Sorm.java
@@ -19,7 +19,7 @@ import org.nkjmlab.sorm4j.internal.mapping.SormImpl;
  */
 public interface Sorm {
 
-  static final SormConfig INITIAL_DEFAULT_CONFIG_STORE = new SormConfigBuilder().build();
+  static final SormConfig DEFAULT_CONFIG = new SormConfigBuilder().build();
 
 
   /**
@@ -29,7 +29,7 @@ public interface Sorm {
    * @return
    */
   static Sorm create(DataSource dataSource) {
-    return Sorm.create(dataSource, INITIAL_DEFAULT_CONFIG_STORE);
+    return Sorm.create(dataSource, DEFAULT_CONFIG);
   }
 
   static Sorm create(DataSource dataSource, SormConfig config) {
@@ -45,7 +45,7 @@ public interface Sorm {
    * @return
    */
   static Sorm create(String jdbcUrl, String user, String password) {
-    return Sorm.create(jdbcUrl, user, password, INITIAL_DEFAULT_CONFIG_STORE);
+    return create(jdbcUrl, user, password, DEFAULT_CONFIG);
   }
 
 
@@ -74,7 +74,7 @@ public interface Sorm {
    * @return
    */
   static OrmConnection toOrmConnection(Connection connection) {
-    return Sorm.toOrmConnection(connection, INITIAL_DEFAULT_CONFIG_STORE);
+    return Sorm.toOrmConnection(connection, DEFAULT_CONFIG);
   }
 
   static OrmConnection toOrmConnection(Connection connection, SormConfig sormConfig) {
