+++ src/main/java/org/nkjmlab/sorm4j/extension/MultiRowProcessorType.java
@@ -1,0 +1,8 @@
+package org.nkjmlab.sorm4j.extension;
+
+/**
+ * Type of how to execute multi-row update SQL statements.
+ */
+public enum MultiRowProcessorType {
+  SIMPLE_BATCH, MULTI_ROW, MULTI_ROW_AND_BATCH
+}
