+++ src/main/java/org/nkjmlab/sorm4j/internal/mapping/OrmImpl.java
@@ -385,5 +385,4 @@ public class OrmImpl implements Orm {
     return new TypedOrmImpl<>(objectClass, this);
   }
 
-
 }
