+++ src/main/java/org/nkjmlab/sorm4j/internal/mapping/multirow/MultiRowProcessorFactory.java
@@ -2,9 +2,9 @@ package org.nkjmlab.sorm4j.internal.mapping.multirow;
 
 import java.util.function.Function;
 import org.nkjmlab.sorm4j.extension.LoggerConfig;
+import org.nkjmlab.sorm4j.extension.MultiRowProcessorType;
 import org.nkjmlab.sorm4j.extension.SormOptions;
 import org.nkjmlab.sorm4j.extension.SqlParametersSetter;
-import org.nkjmlab.sorm4j.extension.SormConfigBuilder.MultiRowProcessorType;
 import org.nkjmlab.sorm4j.internal.mapping.TableMapping;
 
 public final class MultiRowProcessorFactory {
