+++ src/test/java/org/nkjmlab/sorm4j/OrmConnectionTest.java
@@ -16,7 +16,7 @@ import org.nkjmlab.sorm4j.common.Location;
 import org.nkjmlab.sorm4j.common.Player;
 import org.nkjmlab.sorm4j.common.SormTestUtils;
 import org.nkjmlab.sorm4j.extension.DefaultColumnFieldMapper;
-import org.nkjmlab.sorm4j.extension.SormConfigBuilder;
+import org.nkjmlab.sorm4j.extension.SormBuilder;
 import org.nkjmlab.sorm4j.internal.mapping.InsertResultImpl;
 import org.nkjmlab.sorm4j.sql.NamedParameterSql;
 import org.nkjmlab.sorm4j.sql.OrderedParameterSql;
@@ -602,8 +602,8 @@ class OrmConnectionTest {
 
 
 
-    Sorm orm = Sorm.create(sorm.getDataSource(), new SormConfigBuilder()
-        .setTransactionIsolationLevel(Connection.TRANSACTION_SERIALIZABLE).build());
+    Sorm orm = new SormBuilder().setDataSource(sorm.getDataSource())
+        .setTransactionIsolationLevel(Connection.TRANSACTION_SERIALIZABLE).build();
 
     orm.acceptTransactionHandler(m -> {
       assertThat(m.getJdbcConnection().getTransactionIsolation())
