+++ src/test/java/org/nkjmlab/sorm4j/typed/TypedOrmImplTest.java
@@ -10,7 +10,6 @@ import org.nkjmlab.sorm4j.Sorm;
 import org.nkjmlab.sorm4j.common.Player;
 import org.nkjmlab.sorm4j.common.SormTestUtils;
 import org.nkjmlab.sorm4j.sql.ParameterizedSql;
-import org.nkjmlab.sorm4j.typed.TypedOrm;
 
 class TypedOrmImplTest {
